TeamMinds/backend/requirements.txt
fastapi
uvicorn
openai
sqlalchemy
pydantic
httpx
python-dotenv
